Changes according to comments and review
